Reformatted a bit and usleep
Thanks Paul! We're working in that direction (reformatting and putting in comments) and the usleep tip was an excellent idea. I kinda knew the sleep was actually not threadsafe, but thought it wouldn't matter much. This is an improvement and keep the ideas coming! Code is still pretty short and does almost everything that we want it to.